Refactor user-related methods into UsersExtensions
Updated using directives in ServicesExtensions.cs to include necessary namespaces for dependency injection. Removed the `GetUserId` and `GetClaimValue` methods from `ServicesExtensions` and moved them to a newly created `UsersExtensions` class. Added a new `IsAuthenticated` method to enhance user authentication checks, throwing a `UserUnknownException` if the user is not authenticated.
