+++ src/MinimalApi.Identity.API/Extensions/ServicesExtensions.cs
@@ -1,6 +1,4 @@
-﻿using System.Security.Claims;
-using System.Security.Principal;
-using Microsoft.Extensions.Configuration;
+﻿using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Options;
 
@@ -55,23 +53,4 @@ public static class ServicesExtensions
     {
         return services.AddRegisterService<TAssembly>(stringEndsWith, ServiceLifetime.Singleton);
     }
-
-    public static string GetUserId(this IPrincipal user)
-    {
-        if (user is not ClaimsPrincipal claimsPrincipal)
-        {
-            throw new ArgumentException("User must be a ClaimsPrincipal", nameof(user));
-        }
-
-        var claimValue = claimsPrincipal.FindFirst(ClaimTypes.NameIdentifier)?.Value;
-
-        if (string.IsNullOrEmpty(claimValue))
-        {
-            throw new InvalidOperationException("Claim value for NameIdentifier is missing or empty.");
-        }
-
-        return claimValue;
-    }
-
-    public static string GetClaimValue(this IPrincipal user, string claimType) => ((ClaimsPrincipal)user).FindFirst(claimType)?.Value!;
 }
